Precompute moved to be an inlined submodule of params
